requirements.txt
streamlit==1.52.2
google-generativeai==0.8.6
python-dotenv==1.2.1
flake8==7.3.0
pytest==9.0.2